ocr test move createJournalOrOcr  to repository
