+++ src/pages/api/v1/company/[companyId]/ocr/index.repository.ts
@@ -1,4 +1,5 @@
 import prisma from "@/client";
+import { ProgressStatus } from "@/constants/account";
 import { STATUS_MESSAGE } from "@/constants/status_code";
 import { IAccountResultStatus } from "@/interfaces/accounting_account";
 import { timestampInSeconds } from "@/lib/utils/common";
@@ -19,6 +20,7 @@ export async function createOrFindCompanyInPrisma(companyId: number) {
           code: 'TEST_OCR',
           name: 'Company Name',
           regional: 'Regional Name',
+          kycStatus: true,
           startDate: nowTimestamp,
           createdAt: nowTimestamp,
           updatedAt: nowTimestamp,
@@ -90,4 +92,26 @@ export async function upsertJournalInPrisma(
   } catch (error) {
     throw new Error(STATUS_MESSAGE.DATABASE_CREATE_FAILED_ERROR);
   }
+}
+
+export async function createJournalAndOcrInPrisma(
+  companyId: number,
+  aichResult: {
+    resultStatus: IAccountResultStatus;
+    imageUrl: string;
+    imageName: string;
+    imageSize: number;
+  }
+): Promise<void> {
+  // ToDo: (20240521 - Murky) companyId 要檢查是否存在該公司
+  // ToDo: (20240521 - Murky) 重複的圖片一直post貌似會越來越多Journal? 目前沒有檢查重複post的狀況
+  // 如果是AICH已經重複的就不建立了
+  if (aichResult.resultStatus.status !== ProgressStatus.IN_PROGRESS) {
+    return;
+  }
+  await prisma.$transaction(async () => {
+    const company = await createOrFindCompanyInPrisma(companyId);
+    const ocrData = await createOcrInPrisma(aichResult);
+    await upsertJournalInPrisma(company.id, aichResult, ocrData.id);
+  });
 }
